\name{AcreYields}
\docType{data}
\alias{AcreYields}
\title{Acreage and Yields on Corn, Soybeans, and Wheat.}
\description{
  yearly observations from 1930 to 1986

 \emph{number of observations} :  57

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(AcreYields)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{ACorn}{acreage of corn}
    \item{ASoy}{acreage of soy}
    \item{AWheat}{acreage of wheat}
    \item{YCorn}{yields of corn}
    \item{YSoy}{yields of soy}
    \item{YWheat}{yields of wheat}
  }
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 13.8, p.541.
}
\keyword{datasets}