\name{Baltimore}
\docType{data}
\alias{Baltimore}
\title{Data on Population Density in the Different Census Tracts in the Baltimore Area 1970}
\description{
  a cross-section from 1970

 \emph{number of observations} :  39

 \emph{observation} :  individual

   \emph{country} :  United States

}
\usage{data(Baltimore)}
\format{A dataframe containing :
  \describe{
    \item{Obs}{observation}
    \item{y}{density of population in the cencus tract}
    \item{x}{distance of the cencus tract from the central business district}
  }
}
\source{
    Lahiri, K.  and  R.  Numrich (1983) \dQuote{An econometric study of the dynamics of urban spatial structure}, \emph{Journal of Urban Economics }, 55-79.
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 5.5.
}
\keyword{datasets}