\name{ChrisJorg}
\docType{data}
\alias{ChrisJorg}
\title{Output, Labor Input and Capital Input for the United  States, 1929-1967}
\description{
  annual observations from 1929 to 1967

 \emph{number of observations} :  39

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(ChrisJorg)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{X}{index of gross national product in constant dollars}
    \item{L1}{labor input index (number of persons adjusted for hours of work and educational level)}
    \item{L2}{persons engaged,}
    \item{K1}{capital input index (capital stock adjusted for rates of utilization}
    \item{K2}{capital stock in constant dollars}
  }
}
\source{
    Christensen, L.R.  and  D.W.  Jorgenson (1970) \dQuote{U.S. real product and real factor input 1929-67}, \emph{Review of Income and Wealth }, march.
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 3.11, p.97, 170.
}
\keyword{datasets}