\name{Food}
\docType{data}
\alias{Food}
\title{Data on Demand for Food in the United States}
\description{
  annual observations from 1922 to 1941

 \emph{number of observations} :  20

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(Food)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{QD}{food consumption per capita}
    \item{PD}{food prices at retail level/cost of living index}
    \item{Y}{disposable incime/cost of living index}
    \item{QS}{food production per capita}
    \item{PS}{prices reveived by farmers for food/cost of living}
    \item{t}{year index}
  }
}
\source{
    Girschick, M.A.  and  T.  Haarelmo (1947) \dQuote{Statistical analysis of the demand for food}, \emph{Econometrica }, april.
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 4.9, p.194.
}
\keyword{datasets}