\name{Gasol}
\docType{data}
\alias{Gasol}
\title{Data on Gasoline Consumption}
\description{
  annual observations from 1947 to 1969

 \emph{number of observations} :  23

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(Gasol)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{K}{miles traveled per car per year}
    \item{C}{numbers of cars (millions)}
    \item{M}{miles per gallon}
    \item{Pg}{retail gasoline price index deflated by consumer price index (1953=100)}
    \item{Pt}{price of public transport (1967=100)}
    \item{Pop}{population (millions)}
    \item{L}{labor force (millions)}
    \item{Y}{per capita disposable income in 1958 prices}
  }
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., 4.8 p.193.
}
\keyword{datasets}