\name{Housing}
\docType{data}
\alias{Housing}
\title{Data on Housing Starts in Canada}
\description{
  quarterly observations from 1954 & 4 to 1982 & 4

 \emph{number of observations} :  92

 \emph{observation} :  Country

   \emph{country} :  Canada

}
\usage{data(Housing)}
\format{A dataframe containing :
  \describe{
    \item{T}{year and quarter }
    \item{HS}{housing starts}
    \item{y}{gross national expenditures (in 1971 dollars)}
    \item{RR}{estimated real interest rate}
  }
}
\note{\dQuote{Table 5, pp. 515-16. Data on HS and y have been rounded to four-digit numbers..}}
\source{
    Davidson, R.  and  J.G.  MacKinnon (1985) \dQuote{Testing linear and log-linear regressions against box-cox}, \emph{Canadian Journal of Economics }, august.
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 4.10 p.195.
}
\keyword{datasets}