\name{Meat}
\docType{data}
\alias{Meat}
\title{Per Capita Consumption and Deflated Prices of Selected Meats}
\description{
  yearly observations from 1948 to 1963

 \emph{number of observations} :  16

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(Meat)}
\format{A dataframe containing :
  \describe{
    \item{year}{date}
    \item{cbeef}{consumption of beef per capita (lb) }
    \item{pbeef}{price of beef per pound (cents)}
    \item{cpork}{consumption of pork per capita (lb) }
    \item{ppork}{price of pork per pound (cents)}
    \item{clamb}{consumption of lamb per capita (lb) }
    \item{plamb}{price of lamb per pound (cents)}
    \item{cveal}{consumption of veal per capita (lb) }
    \item{pveal}{price of veal per pound (cents)}
    \item{cchicken}{consumption of chicken per capita (lb) }
    \item{pchicken}{price of chicken per pound (cents)}
  }
}
\note{\dQuote{1963 data are preliminary and were not used in the analysis, consumption is in carcass weight equivalent, prices are divided by consumer price index (1957-1959=100).}}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd.,  Table 8.2, p.310 	.
}
\keyword{datasets}