\name{NelsPlos}
\docType{data}
\alias{NelsPlos}
\title{Nelson and Plosser's Data}
\description{
  yearly observations from 1860 to 1970

 \emph{number of observations} :  111

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(NelsPlos)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{RGNP}{real GNP}
    \item{GNP}{nominal GNP}
    \item{PCRGNP}{real per capita GNP}
    \item{IP}{industrial production}
    \item{EMP}{employment}
    \item{UN}{unemployment rate}
    \item{PRGNP}{gNP deflator}
    \item{CPI}{consumer price}
    \item{WG}{wages}
    \item{RWG}{real wages}
    \item{M}{money stock}
    \item{VEL}{velocity}
    \item{BND}{bound yield}
    \item{SP500}{common stock price}
  }
}
\source{
    Nelson,  and    Plosser (1982) \dQuote{trends and random walks in macroeconomic series; some evidence and implications}, \emph{Journal of Monetary Economics }, 139-162.
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table13.3, p.538.
}
\keyword{datasets}