\name{Pork}
\docType{data}
\alias{Pork}
\title{Demand and Supply of Pork, United States, 1922-1941}
\description{
  yearly observations from 1922 to 1941

 \emph{number of observations} :  20

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(Pork)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{Pt}{retail price of pork (cents per pound)}
    \item{Qt}{consumption of pork (pounds per capita)}
    \item{Yt}{disposable personal income (dollars per capita)}
    \item{Zt}{predetermined elements in pork production}
  }
}
\note{\dQuote{p.539.}}
\source{
    Merril, William C.  and  Karl A.  Fox (1971) \emph{Introduction to economic statistics}, New York: Wiley.
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 9.1 	.
}
\keyword{datasets}