\name{QConsInc}
\docType{data}
\alias{QConsInc}
\title{Quarterly Data on Consumption and Income}
\description{
  quarterly observations from 1947-1 to 1981-2

 \emph{number of observations} :  138

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(QConsInc)}
\format{A dataframe containing :
  \describe{
    \item{Year}{year}
    \item{Quarter}{quarter}
    \item{Consumption}{real per capita consumption on non durables}
    \item{Income}{real per capita disposable income}
  }
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 13.4, p.541.
}
\keyword{datasets}