\name{QGNPM}
\docType{data}
\alias{QGNPM}
\title{Quarterly Data on GNP and Money}
\description{
  quarterly observations from 59-1 to 81-1

 \emph{number of observations} :  89

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(QGNPM)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{Quarter}{quarter}
    \item{GNP}{nominal GNP}
    \item{QM2}{money supply measured by M2 }
  }
}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., 13.5.
}
\keyword{datasets}