\name{RuralLand}
\docType{data}
\alias{RuralLand}
\title{ Sales Price of Rural Land}
\description{
  quarterly observations from 59-1 to 81-1

 \emph{number of observations} :  67

 \emph{observation} :  Individual

   \emph{country} :  United States

}
\usage{data(RuralLand)}
\format{A dataframe containing :
  \describe{
    \item{N}{index}
    \item{Price}{price}
    \item{WL}{proportion of acreage that is wooded}
    \item{DA}{distance form parcel to Sarasota airport}
    \item{D75}{distance from parcel to I-75}
    \item{A}{acreage of parcel}
    \item{MO}{moutn in which the parcel was sold}
  }
}
\note{\dQuote{Price, observed land price per acre excluding improvements.}}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 4.7 p.193.
}
\keyword{datasets}