\name{StockPr}
\docType{data}
\alias{StockPr}
\title{Real Stock Price (S and P) and Real Dividends (S and P), 1871-1985}
\description{
  yearly observations from 1871 to 1985

 \emph{number of observations} :  115

 \emph{observation} :  Country

   \emph{country} :  United States

}
\usage{data(StockPr)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date}
    \item{Price}{real stock price}
    \item{Dividends}{real dividends}
  }
}
\note{\dQuote{Stock prices and dividends corresponding to the Standard and Poor's 500 index.}}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., 13.6.
}
\keyword{datasets}