\name{Wages}
\docType{data}
\alias{Wages}
\title{Wages, Benefits, Unemployment and Net National Product: UK 1920-1938}
\description{
  annual observations from 1920 to 1938

 \emph{number of observations} :  19

 \emph{observation} :  Country

   \emph{country} :  United Kingdown

}
\usage{data(Wages)}
\format{A dataframe containing :
  \describe{
    \item{Year}{date 	}
    \item{W}{weekly wages, in shilling}
    \item{B}{weekly benefits, in shilling}
    \item{U}{unemployment rate (\%)}
    \item{BW}{benefits/wages}
    \item{Q}{net national product (millions of pounds  at 1938 factor cost)}
  }
}
\note{\dQuote{shilling=1/20 pond.}}
\references{
    Maddala, G.S. (2001) \emph{Introduction to Econometrics}, John Wiley and sons, ltd., Table 4.11, p.196.
}
\keyword{datasets}